# openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beyond Auth",
    "description": "Per-project authentication and authorization service.",
    "license": {
      "name": "AGPL-3.0",
      "identifier": "AGPL-3.0"
    },
    "version": "1"
  },
  "paths": {
...
    "/v1/keys": {
      "get": {
        "tags": [
          "keys"
        ],
        "summary": "List all API keys for the authenticated user. Bearer token values are never returned\n— only metadata (ID, name, created/used/expires timestamps).",
        "operationId": "list_keys",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeysResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "keys"
        ],
        "summary": "Create an API key for the authenticated user. The full bearer token is returned in\n`key` and is shown only once — store it immediately. Subsequent reads return only metadata.",
        "operationId": "create_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Key created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/keys/{id}": {
      "get": {
        "tags": [
          "keys"
        ],
        "summary": "Get metadata for a single API key.",
        "operationId": "get_key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Key ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Key"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "keys"
        ],
        "summary": "Revoke an API key. The key is immediately invalid for authentication.",
        "operationId": "delete_key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Key ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Key revoked"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the key expires. Defaults to 100 years from now (effectively never)."
          },
          "name": {
            "type": "string"
          }
        }
      },
      "CreateResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Key"
          },
          {
            "type": "object",
            "required": [
              "key"
            ],
            "properties": {
              "key": {
                "type": "string",
                "description": "The full bearer token. Store it now — it cannot be retrieved later."
              }
            }
          }
        ],
        "description": "Returned only on creation — the `key` field is never shown again."
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Wire-format error body returned on all non-2xx responses.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code, e.g. `\"invalid_credentials\"`."
          },
          "hint": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional actionable guidance present on configuration-gate errors."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Top-level error envelope for all API error responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "Key": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "expires_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_used_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "KeysResponse": {
        "type": "object",
        "required": [
          "keys"
        ],
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Key"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "session_<id>_<secret>"
      }
    }
  },
...
}